[artf494694, artf494700, artf494701] - System populated items should not be deleted, Changes in unit Test, Changes in Customers, field empty validation modified

+++ service/phresco-service-web/src/main/java/com/photon/phresco/service/admin/actions/admin/Customers.java
@@ -35,6 +35,7 @@ public class Customers extends ServiceBaseAction  {
 	
 	private static final long serialVersionUID = 6801037145464060759L;
 	private static final Logger S_LOGGER = Logger.getLogger(Customers.class);
+	private static Boolean isDebugEnabled = S_LOGGER.isDebugEnabled();
 	
 	private String name = null;
 	private String nameError = null;
@@ -51,7 +52,8 @@ public class Customers extends ServiceBaseAction  {
 	private String faxError = null;
 	private String country = null;
 	private String conError = null;
-	private String licence = null;
+	private String state = null;
+    private String licence = null;
 	private String licenError = null;
 	private boolean errorFound = false;
 	private String validFrom = null;
@@ -60,9 +62,12 @@ public class Customers extends ServiceBaseAction  {
 	private String fromPage = null;
 	private String customerId = null;
 	private String oldName = null;
+	private String helpText = null;
 
-	public String list() throws PhrescoException {
-		S_LOGGER.debug("Entering Method  CustomerList.list()");
+    public String list() throws PhrescoException {
+	    if (isDebugEnabled) {
+	        S_LOGGER.debug("Entering Method CustomerList.list()");
+	    }
 		
 		try {
             List<Customer> customers = getServiceManager().getCustomers();
@@ -75,7 +80,9 @@ public class Customers extends ServiceBaseAction  {
 	}
 	
 	public String add() throws PhrescoException {
-		S_LOGGER.debug("Entering Method  CustomerList.add()");
+	    if (isDebugEnabled) {
+	        S_LOGGER.debug("Entering Method CustomerList.add()");
+	    }
 		
 		try {
 			if (REQ_EDIT.equals(fromPage)) {
@@ -91,15 +98,22 @@ public class Customers extends ServiceBaseAction  {
 	}
 	
 	public String save() throws PhrescoException {
-		S_LOGGER.debug("Entering Method  CustomerList.save()");
+	    if (isDebugEnabled) {
+	        S_LOGGER.debug("Entering Method CustomerList.save()");
+	    }
+	    
 		try {
-			if (validateForm()) {
-				setErrorFound(true);
-				
-				return SUCCESS;
-			}
 			List<Customer> customers = new ArrayList<Customer>();
 			Customer customer = new Customer(name, description);
+			customer.setEmailId(email);
+			customer.setAddress(address);
+			customer.setCountry(country);
+			customer.setState(state);
+			customer.setZipcode(zipcode);
+			customer.setContactNumber(number);
+			customer.setFax(fax);
+			customer.setHelpText(helpText);
+			customer.setType(Integer.parseInt(licence));
 			customer.setValidFrom(validFrom);
 			customer.setValidUpto(validUpTo);
 			customer.setRepoURL(repoURL);
@@ -118,18 +132,24 @@ public class Customers extends ServiceBaseAction  {
 	}
 
 	public String update() throws PhrescoException {
-		S_LOGGER.debug("Entering Method  CustomerList.update()");
+	    if (isDebugEnabled) {
+	        S_LOGGER.debug("Entering Method CustomerList.update()");
+	    }
 
 		try {
-			if (validateForm()) {
-				setErrorFound(true);
-				
-				return SUCCESS;
-			}
 			Customer customer = new Customer(customerId, name, description);
-			customer.setValidFrom(validFrom);
-			customer.setValidUpto(validUpTo);
-			customer.setRepoURL(repoURL);
+			customer.setEmailId(email);
+            customer.setAddress(address);
+            customer.setCountry(country);
+            customer.setState(state);
+            customer.setZipcode(zipcode);
+            customer.setContactNumber(number);
+            customer.setFax(fax);
+            customer.setHelpText(helpText);
+            customer.setType(Integer.parseInt(licence));
+            customer.setValidFrom(validFrom);
+            customer.setValidUpto(validUpTo);
+            customer.setRepoURL(repoURL);
 			getServiceManager().updateCustomer(customer, customerId);
 		} catch(Exception e)  {
 			throw new PhrescoException(e);
@@ -138,7 +158,11 @@ public class Customers extends ServiceBaseAction  {
 		return list();
 	}
 
-	private boolean validateForm() {
+	public String validateForm() {
+	    if (isDebugEnabled) {
+	        S_LOGGER.debug("Entering Method CustomerList.validateForm()");
+        }
+	    
 		boolean isError = false;
 		if (StringUtils.isEmpty(name)) {
 			setNameError(getText(KEY_I18N_ERR_NAME_EMPTY));
@@ -180,11 +204,17 @@ public class Customers extends ServiceBaseAction  {
 			isError = true;
 		}
 		
-		return isError;
+		if (isError) {
+            setErrorFound(true);
+        }
+		
+		return SUCCESS;
 	}
 	
 	public String delete() throws PhrescoException {
-		S_LOGGER.debug("Entering Method  CustomerList.delete()");
+	    if (isDebugEnabled) {
+	        S_LOGGER.debug("Entering Method CustomerList.delete()");
+	    }
 		
 		try {
 			String[] customerIds = getHttpRequest().getParameterValues(REQ_CUST_CUSTOMER_ID);
@@ -395,4 +425,20 @@ public class Customers extends ServiceBaseAction  {
 	public void setOldName(String oldName) {
 		this.oldName = oldName;
 	}
+	
+	public String getState() {
+        return state;
+    }
+
+    public void setState(String state) {
+        this.state = state;
+    }
+    
+    public String getHelpText() {
+        return helpText;
+    }
+
+    public void setHelpText(String helpText) {
+        this.helpText = helpText;
+    }
 }
